Update for MemoryError.

+++ Deep-Learning-by-means-of-Design-Pattern/pydbm/synapse/recurrenttemporalgraph/lstm_graph.pyx
@@ -453,12 +453,12 @@ class LSTMGraph(RecurrentTemporalGraph):
         self.hidden_activity_arr = np.array([])
         self.rnn_activity_arr = np.array([])
 
-        self.weights_lstm_observed_arr = np.random.normal(size=(input_neuron_count, hidden_neuron_count * 4)) * 0.1
-        self.weights_lstm_hidden_arr = np.random.normal(size=(hidden_neuron_count, hidden_neuron_count * 4)) * 0.1
-        self.lstm_bias_arr = np.zeros(hidden_neuron_count * 4)
+        self.weights_lstm_observed_arr = np.random.normal(size=(input_neuron_count, hidden_neuron_count * 4)).astype(np.float16) * 0.1
+        self.weights_lstm_hidden_arr = np.random.normal(size=(hidden_neuron_count, hidden_neuron_count * 4)).astype(np.float16) * 0.1
+        self.lstm_bias_arr = np.zeros(hidden_neuron_count * 4).astype(np.float16)
 
-        self.weights_output_arr = np.random.normal(size=(hidden_neuron_count, output_neuron_count)) * 0.1
-        self.output_bias_arr = np.zeros(output_neuron_count)
+        self.weights_output_arr = np.random.normal(size=(hidden_neuron_count, output_neuron_count)).astype(np.float16) * 0.1
+        self.output_bias_arr = np.zeros(output_neuron_count).astype(np.float16)
 
     def create_node(
         self,
